Low_End_Web_App/requirements.txt
streamlit>=1.28.0
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0
opencv-python>=4.7.0
pandas>=1.4.0
tqdm>=4.64.0
plotly>=5.15.0
